+++ ref-exchange/tests/test_migrate.rs
@@ -9,7 +9,7 @@ use crate::common::utils::*;
 pub mod common;
 
 near_sdk_sim::lazy_static_include::lazy_static_include_bytes! {
-    PREV_EXCHANGE_WASM_BYTES => "../releases/ref_exchange_release_v172.wasm",
+    PREV_EXCHANGE_WASM_BYTES => "../releases/ref_exchange_release_v180.wasm",
     EXCHANGE_WASM_BYTES => "../res/ref_exchange.wasm",
 }
 
@@ -49,10 +49,10 @@ fn test_upgrade() {
     .assert_success();
     let metadata = get_metadata(&pool);
     // println!("{:#?}", metadata);
-    assert_eq!(metadata.version, "1.8.0".to_string());
+    assert_eq!(metadata.version, "1.9.0".to_string());
     assert_eq!(metadata.admin_fee_bps, 5);
-    assert_eq!(metadata.boost_farm_id, root.account_id());
-    assert_eq!(metadata.burrowland_id, root.account_id());
+    assert_eq!(metadata.boost_farm_id, "boost_farm".to_string());
+    assert_eq!(metadata.burrowland_id, "burrowland".to_string());
     assert_eq!(metadata.state, RunningState::Running);
 
     // Upgrade to the same code with insurfficient gas.